refactor h5peditor/ h5pplayer service to provider
Hint: npm ci
