+++ apps/server/src/modules/h5p-editor/provider/h5p-editor.provider.ts
@@ -1,13 +1,13 @@
 import { H5PEditor, cacheImplementations } from '@lumieducation/h5p-server';
 
 import { IH5PEditorOptions, ITranslationFunction } from '@lumieducation/h5p-server/build/src/types';
-import { h5pConfig, h5pUrlGenerator } from './config/h5p-service-config';
-import { ContentStorage } from './contentStorage.service';
-import { Translator } from './h5p-translator.service';
-import { LibraryStorage } from './libraryStorage.service';
-import { TemporaryFileStorage } from './temporary-file-storage.service';
+import { h5pConfig, h5pUrlGenerator } from '../service/config/h5p-service-config';
+import { ContentStorage } from '../service/contentStorage.service';
+import { Translator } from '../service/h5p-translator.service';
+import { LibraryStorage } from '../service/libraryStorage.service';
+import { TemporaryFileStorage } from '../service/temporary-file-storage.service';
 
-export const H5PEditorService = {
+export const H5PEditorProvider = {
 	provide: H5PEditor,
 	inject: [ContentStorage, LibraryStorage, TemporaryFileStorage],
 	async useFactory(
