+++ apps/server/src/modules/h5p-editor/provider/h5p-player.provider.ts
@@ -1,11 +1,11 @@
 import { H5PPlayer, ITranslationFunction } from '@lumieducation/h5p-server';
 
-import { h5pConfig, h5pUrlGenerator } from './config/h5p-service-config';
-import { ContentStorage } from './contentStorage.service';
-import { Translator } from './h5p-translator.service';
-import { LibraryStorage } from './libraryStorage.service';
+import { h5pConfig, h5pUrlGenerator } from '../service/config/h5p-service-config';
+import { ContentStorage } from '../service/contentStorage.service';
+import { Translator } from '../service/h5p-translator.service';
+import { LibraryStorage } from '../service/libraryStorage.service';
 
-export const H5PPlayerService = {
+export const H5PPlayerProvider = {
 	provide: H5PPlayer,
 	inject: [ContentStorage, LibraryStorage],
 	useFactory: async (contentStorage: ContentStorage, libraryStorage: LibraryStorage) => {
