+++ apps/server/src/modules/h5p-editor/provider/index.ts
@@ -1,0 +1,2 @@
+export * from './h5p-editor.provider';
+export * from './h5p-player.provider';
