+++ apps/server/src/modules/h5p-editor/service/index.ts
@@ -1,5 +1,3 @@
-export * from './h5p-editor.service';
-export * from './h5p-player.service';
 export * from './h5p-ajax-endpoint.service';
 export * from './contentStorage.service';
 export * from './libraryStorage.service';
